+++ datasketches/tests/cpc_test/main.rs
@@ -15,8 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! Integration tests for CPC sketches and set operations.
-
 mod union;
 mod update;
 mod wrapper;
