+++ datasketches/tests/hll_test/main.rs
@@ -15,7 +15,5 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! Integration tests for HLL sketches and unions.
-
 mod union;
 mod update;
