+++ datasketches/tests/serde_tests/hll.rs
@@ -22,7 +22,7 @@ use datasketches::hash_value::natural_extend;
 use datasketches::hll::HllSketch;
 use datasketches::hll::HllType;
 
-use crate::support::serialization_test_data;
+use crate::serialization_test_data;
 
 fn test_sketch_file(path: PathBuf, expected_cardinality: usize, expected_lg_k: u8) {
     let expected = expected_cardinality as f64;
