+++ datasketches/tests/serde_tests/tdigest.rs
@@ -23,7 +23,7 @@ use googletest::assert_that;
 use googletest::prelude::eq;
 use googletest::prelude::near;
 
-use crate::support::serialization_test_data;
+use crate::serialization_test_data;
 
 fn test_sketch_file(path: PathBuf, n: u64, with_buffer: bool, is_f32: bool) {
     let bytes = fs::read(&path).unwrap();
