+++ datasketches/tests/tuple_test/main.rs
@@ -15,8 +15,6 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! Integration tests for tuple sketches and set operations.
-
 mod a_not_b;
 mod intersection;
 mod sketch;
